# public/openapi.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Empresa de Transportes El Cumbe - API RESTful",
                 "version":  "1.0.0",
                 "description":  "EspecificaciÃ³n oficial de servicios RESTful y API MÃ³vil para la plataforma de transporte interprovincial El Cumbe. Incluye servicios de autenticaciÃ³n JWT, bÃºsqueda de salidas, reserva de asientos, pasarela de pagos Culqi y gestiÃ³n de encomiendas."
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3000",
                        "description":  "Servidor Local de Desarrollo"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT",
                                                                  "description":  "Ingrese el token JWT obtenido en la API de Login MÃ³vil (/api/movil/login)"
                                                              }
                                           },
                       "schemas":  {
...
                                       "RegistroRequest":  {
                                                               "type":  "object",
                                                               "required":  [
                                                                                "nombres",
                                                                                "apellidos",
                                                                                "correo",
                                                                                "contrasena",
                                                                                "dni"
                                                                            ],
                                                               "properties":  {
                                                                                  "nombres":  {
                                                                                                  "type":  "string",
                                                                                                  "example":  "Juan Carlos"
                                                                                              },
                                                                                  "apellidos":  {
                                                                                                    "type":  "string",
                                                                                                    "example":  "PÃ©rez Quispe"
                                                                                                },
                                                                                  "correo":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "email",
                                                                                                 "example":  "juan.perez@gmail.com"
                                                                                             },
                                                                                  "contrasena":  {
                                                                                                     "type":  "string",
                                                                                                     "example":  "123456"
                                                                                                 },
                                                                                  "dni":  {
                                                                                              "type":  "string",
                                                                                              "example":  "74829104"
                                                                                          },
                                                                                  "telefono":  {
                                                                                                   "type":  "string",
                                                                                                   "example":  "918273645"
                                                                                               }
                                                                              }
                                                           },
...
                                   }
                   },
    "paths":  {
...
                  "/api/movil/registro":  {
                                              "post":  {
                                                           "summary":  "Registro de Clientes (MÃ³vil)",
                                                           "description":  "Crea una cuenta de usuario con rol \u0027cliente\u0027 e inserta sus datos en la entidad Persona con validaciÃ³n de DNI Ãºnico.",
                                                           "requestBody":  {
                                                                               "required":  true,
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/RegistroRequest"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Registro exitoso"
                                                                                     },
                                                                             "400":  {
                                                                                         "description":  "Error de validaciÃ³n o usuario existente"
                                                                                     }
                                                                         }
                                                       }
                                          },
...
                  "/api/movil/pasajes/validar-qr":  {
                                                        "post":  {
                                                                     "tags":  [
                                                                                  "Operario"
                                                                              ],
                                                                     "summary":  "Validar QR y registrar abordaje de forma atómica",
                                                                     "security":  [
                                                                                      {
                                                                                          "bearerAuth":  [

                                                                                                         ]
                                                                                      }
                                                                                  ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "Operación exitosa"
                                                                                               },
                                                                                       "400":  {
                                                                                                   "description":  "Solicitud inválida"
                                                                                               },
                                                                                       "401":  {
                                                                                                   "description":  "Token ausente o inválido"
                                                                                               },
                                                                                       "403":  {
                                                                                                   "description":  "Rol o recurso no autorizado"
                                                                                               },
                                                                                       "500":  {
                                                                                                   "description":  "Error interno"
                                                                                               }
                                                                                   },
                                                                     "requestBody":  {
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "type":  "object"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     },
                                                                                         "required":  true
                                                                                     }
                                                                 }
                                                    }
              }
}